+++ lib/src/endpoint.dart
@@ -7,5 +7,6 @@ class Endpoint {
 
   String get suggestAddress => '$baseUrl/suggest/address';
 
-  String get geolocationAddress => '$baseUrl/geolocate/address';
+  String get geolocateAddress => '$baseUrl/geolocate/address';
+  String get passportIssuedBy => '$baseUrl/suggest/fms_unit';
 }