+++ lib/src/model/common/suggestion_request.dart
@@ -1,0 +1,7 @@
+abstract class SuggestionRequest {
+  /// Request query
+  String get query;
+
+  /// Suggestions response count (default 10, max 20)
+  int? get count;
+}